updated the database for products, I'm still working through the seeding.
