+++ server/models/Products.js
@@ -1,7 +1,27 @@
 const mongoose = require('mongoose');
 
+//Added the review schema
+const reviewSchema = new mongoose.Schema({
+  username: {
+    type: String,
+    required: true,
+  },
+  rating: {
+    type: Number,
+    required: true,
+  },
+  text: {
+    type: String,
+  },
+  date: {
+    type: Date,
+    default: Date.now,
+  },
+});
+
 const productSchema = new mongoose.Schema({
-  productType: {
+  //updated to name
+  name: {
     type: String,
     required: true,
   },
@@ -9,13 +29,13 @@ const productSchema = new mongoose.Schema({
     type: String,
     required: true,
   },
-  size: {
-    type: String,
+  sizes: {
+    type: [String],
     enum: ['small', 'medium', 'large'],
     required: true,
   },
-  color: {
-    type: String,
+  colors: {
+    type: [String],
     enum: ['yellow', 'black', 'purple', 'pink', 'blue'],
     required: true,
   },
@@ -23,14 +43,21 @@ const productSchema = new mongoose.Schema({
     type: String,
     required: true,
   },
-  threedModel: { 
+  //changed name to model
+  model: {
     type: String,
-    required: false,
   },
   price: {
     type: Number,
     required: true,
   },
+  reviews: [reviewSchema],
 });
 
-module.exports = mongoose.model('Product', productSchema);
+//updated the export to be be an object. if we split this into 2 schemas the object will no longer need to be destructerd in  the index.
+module.exports = {
+  Product: mongoose.model('Product', productSchema),
+  Review: mongoose.model('Review', reviewSchema)
+};
+
+
